A2 00012506 未命名 南昌工程学院/app/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
PyMySQL==1.1.0
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0
redis==5.0.1
Werkzeug==2.3.7
email-validator==2.0.0 